add Note Listener to seperate class again
