+++ samples/SoundBoard/src/main/java/com/google/oboe/samples/soundboard/MainActivity.java
@@ -1,7 +1,7 @@
 package com.google.oboe.samples.soundboard;
 
 /*
- * Copyright 2018 The Android Open Source Project
+ * Copyright 2021 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -106,7 +106,7 @@ public class MainActivity extends AppCompatActivity {
             }
         }
         
-        setContentView(new MusicTileView(this, rectangles, mEngineHandle));
+        setContentView(new MusicTileView(this, rectangles, new NoteListener(mEngineHandle)));
     }
 }
