+++ samples/SoundBoard/src/main/java/com/google/oboe/samples/soundboard/NoteListener.java
@@ -1,0 +1,36 @@
+package com.google.oboe.samples.soundboard;
+
+/*
+ * Copyright 2021 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+public class NoteListener implements MusicTileView.TileListener {
+    private native void noteOn(long engineHandle, int noteIndex);
+    private native void noteOff(long engineHandle, int noteIndex);
+
+    long mEngineHandle;
+
+    public NoteListener(long engineHandle) {
+        mEngineHandle = engineHandle;
+    }
+
+    public void onTileOn(int index) {
+        noteOn(mEngineHandle, index);
+    }
+
+    public void onTileOff(int index) {
+        noteOff(mEngineHandle, index);
+    }
+}
